% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_defined}
\alias{is_defined}
\title{Test existance of a variable}
\usage{
is_defined(name, envir = .GlobalEnv)
}
\arguments{
\item{name}{the name of the object to search for}

\item{envir}{the environment to search in; defaults to global}
}
\description{
A utility function that tests whether an object has been defined by the given name
 in the specified environment.
}
\examples{
x <- 42
is_defined("x") #TRUE

}
