% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_model.R
\name{Question}
\alias{Question}
\title{Question constructor}
\usage{
Question(q_text, q_tester, q_type = "console-input", hints = NULL)
}
\arguments{
\item{q_text}{the text that will be displayed in the console prompting the user
to action.}

\item{q_tester}{an instance of class \link[bintestr]{QuestionTester}.}

\item{q_type}{the type can be one of \code{console-input}, \code{environment-eval}. For
\code{console-input}, execution will be halted until the user has entered some text
and hit enter. For \code{environment-eval}, the user is required to explicitly call
\code{verify()} on the question instance.}

\item{hints}{a list of hints that will be displayed in consecutive order for each
incorrect answer.}
}
\description{
This S3 class encapsulates a question for interactive lesson scenarios.
}
\examples{
## Instantiate a QuestionTester that tests for a character vector with a single
## "word" of length 5
## The feedback_generator generates some appropriate messages depending on the answer
t1 <- QuestionTester(
    function(x){is.character(x) && length(x)==1 && nchar(x)==5},
    function(x){
        if(! is.character(x)) "that is no word"
        else if(nchar(x) < 5) paste0(x, " has ", nchar(x), " characters and that is not 5")
    }
)
q1 <- Question("Please give a word of length 5", t1)

}
