% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{start_lesson}
\alias{start_lesson}
\title{loads the lesson registered as current}
\usage{
start_lesson(lesson = "__current__", question = "__current__")
}
\description{
Loads the current lesson (or the one specified as argument)
}
