% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{instructions}
\alias{instructions}
\title{Display help (instructions)}
\usage{
instructions()
}
\description{
- \code{start()} - To proceed where you left last time (or at question 1 if
 you just started).
}
\details{
- \code{check.me()} - Will verify your current question, and maybe give some
feedback if it is not correct

- \code{next.question()} - Will proceed to the next question
}
