% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{instructions}
\alias{instructions}
\title{Display bintestr instructions}
\usage{
instructions()
}
\description{
This help page lists the functions you can use to navigate this course.
}
\details{
- \code{start(question="first")} - To proceed where you left last time (or at question 1 if
 you just started).

- \code{check.me()} - Will verify your current question, and maybe give some
feedback if it is not correct

- \code{next_question()} - Will proceed to the next question

- \code{bye()} - Saves your progress and quits - but you will not want to quit!
}
